so gadu now compiles fully :D
svn path=/trunk/KDE/kdenetwork/kopete/; revision=453858

+++ protocols/gadu/gadurichtextformat.cpp
@@ -61,7 +61,7 @@ GaduRichTextFormat::convertToHtml( const QString& msg, unsigned int formats, voi
 		char font = format->font;
 		QString style;
 
-		if ( position < j || position > msg.length() ) {
+		if ( position < j || position > (unsigned int)msg.length() ) {
 			break;
 		}
